+++ src/types/counter/unix_time/constants/seconds.rs
