+++ src/types/mod.rs
@@ -27,5 +27,5 @@
  */
 
 pub(crate) mod data;
-pub(crate) mod it;
+pub(crate) mod counter;
 pub(crate) mod planets;
