+++ src/types/planets/earth/calendar/functions/mod.rs
@@ -26,6 +26,8 @@
  * THE SOFTWARE.
  */
 
+mod era_days;
 mod leap_year;
 
+pub use era_days::{era_days_from_date};
 pub use leap_year::{is_leap_year, sum_leap_years};
